+++ pages/helpers/cdc_abmclientedb.php
@@ -10,6 +10,8 @@
     $cuit = $_POST['cuit'];
     $sector = $_POST['sector'];
     $convenio = $_POST['convenio'];
+    $ejecutivo_cuenta = $_POST['ejecutivo_cuenta'];
+    $con_servicio_correo = $_POST['servicio_correo'];
 
     $result = new stdClass();
     $result->ok = false;
@@ -17,8 +19,8 @@
     switch ($op) {
         case 'A':
             // INSERT
-            $insert_cliente = mysqli_query($con, "INSERT INTO cdc_cliente(id_organismo, razon_social, nombre_corto,cuit, sector, con_convenio) 
-                                                VALUES ('$id_organismo', '$razon_social', '$nombre_corto','$cuit', '$sector', '$convenio')") or die(mysqli_error());	
+            $insert_cliente = mysqli_query($con, "INSERT INTO cdc_cliente(id_organismo, razon_social, nombre_corto,cuit, sector, con_convenio, con_servicio_correo, ejecutivo_cuenta) 
+                                                VALUES ('$id_organismo', '$razon_social', '$nombre_corto','$cuit', '$sector', '$convenio', '$con_servicio_correo', '$ejecutivo_cuenta')") or die(mysqli_error());	
             $lastInsert = mysqli_insert_id($con);
             $insert_audit = mysqli_query($con, "INSERT INTO auditoria (evento, item, id_item, fecha, usuario) 
                                                 VALUES ('1', '2', '$lastInsert', now(), '$user')") or die(mysqli_error());
@@ -27,7 +29,7 @@
         
         case 'M':
             //UPDATE
-            $update_clietne = mysqli_query($con, "UPDATE cdc_cliente SET id_organismo='$id_organismo', razon_social='$razon_social', nombre_corto='$nombre_corto', cuit='$cuit', sector='$sector' , con_convenio='$convenio'
+            $update_clietne = mysqli_query($con, "UPDATE cdc_cliente SET id_organismo='$id_organismo', razon_social='$razon_social', nombre_corto='$nombre_corto', cuit='$cuit', sector='$sector' , con_convenio='$convenio', con_servicio_correo='$con_servicio_correo', ejecutivo_cuenta='$ejecutivo_cuenta'
                                                     WHERE id='$id'") or die(mysqli_error());	
             $insert_audit = mysqli_query($con, "INSERT INTO auditoria (evento, item, id_item, fecha, usuario) 
                                                 VALUES ('1', '2', '$lastInsert', now(), '$user')") or die(mysqli_error());