+++ pages/modals/cdc_clientes_baja_view.php
@@ -1,0 +1,46 @@
+<div class="modal fade" id="modal-abm-clientes-baja-view">
+    <div class="modal-dialog" style="width:1500px;">
+    <div class="modal-content">
+        <div class="modal-header">
+            <button type="button" class="close" data-dismiss="modal" aria-label="Close"><span aria-hidden="true">&times;</span></button>
+            <h2 class="modal-title" id='modal-abm-clientes-baja-view-title'>Clientes dados de baja</h2>
+            
+        </div>
+        <div class="modal-body">
+                
+                <div class="box-body">
+                    <div class="row">
+                        <div class="col-xs-12">
+                            <div class="box">
+                                <div class="box-header">
+                                    <div class="col-sm-6" style="text-align:left">
+                                        <h2 class="box-title">Listados de clientes dados de baja</h2>
+                                    </div>
+                                </div>
+
+                                <!-- /.box-header -->
+                                <div class="box-body">
+                                    <table id="tbbaja" class="display" width="100%">
+                                        <thead>
+                                            <tr>
+                                            <th>Organismo</th>
+                                            <th>Cliente</th>
+                                            <th>Alias/Sigla</th>
+                                            <th>CUIT</th>
+                                            <th>Sector</th>
+                                            </tr>
+                                        </thead>
+                                    </table>
+                                </div>
+                                <!-- /.box-body -->
+                            </div>
+                        </div>
+                        <!-- /.col -->
+                    </div>
+                </div>
+        </div>
+    </div>
+        <!-- /.modal-content -->
+    </div>
+    <!-- /.modal-dialog -->
+</div>