+++ pages/site_sidebar.php
@@ -114,6 +114,7 @@
             <li><a href="sdc_housing.php"><i class="fa fa-home"></i> Housing</a></li>
             <li><a href="sdc_hosting.php"><i class="fa fa-server"></i> Hosting</a></li>
             <li><a href="sdc_iaas.php"><i class="fa fa-cloud"></i> IAAS</a></li>
+            <li><a href="sdc_correo.php"><i class="fa fa-envelope"></i> Correo</a></li>
           </ul>
         </li>
       </ul>