+++ site.php
@@ -455,6 +455,7 @@ desired effect
                     <li><a href="./pages/sdc_housing.php"><i class="fa fa-home"></i> Housing</a></li>
                     <li><a href="./pages/sdc_hosting.php"><i class="fa fa-server"></i> Hosting</a></li>
                     <li><a href="./pages/sdc_iaas.php"><i class="fa fa-cloud"></i> IAAS</a></li>
+                    <li><a href="./pages/sdc_correo.php"><i class="fa fa-envelope"></i> Correo</a></li>
                   </ul>
                 </li>
               </ul>